+++ src/app/componentRegistry.js
@@ -7,6 +7,9 @@ import SpatialSubscriber from '../components/spatial/SpatialSubscriber';
 import FactorsSubscriber from '../components/factors/FactorsSubscriber';
 import GenesSubscriber from '../components/genes/GenesSubscriber';
 import HeatmapSubscriber from '../components/heatmap/HeatmapSubscriber';
+import HoverableScatterplotSubscriber from '../components/scatterplot/HoverableScatterplotSubscriber';
+import HoverableSpatialSubscriber from '../components/spatial/HoverableSpatialSubscriber';
+import HoverableHeatmapSubscriber from '../components/heatmap/HoverableHeatmapSubscriber';
 
 class Description extends React.Component {
   componentDidMount() {
@@ -32,6 +35,9 @@ const registry = {
   FactorsSubscriber,
   GenesSubscriber,
   HeatmapSubscriber,
+  HoverableScatterplotSubscriber,
+  HoverableSpatialSubscriber,
+  HoverableHeatmapSubscriber,
   HiGlassComponent: React.lazy(() => import('../vendor/StyledHiGlass.js')),
   HiGlassWrappedComponent: React.lazy(() => import('../vendor/WrappedStyledHiGlass.js')),
   // TODO: Our Higlass demos don't actually need any data layers to load...