+++ src/components/heatmap/Heatmap.js
@@ -5,7 +5,18 @@ import HeatmapCellColorCanvas from './HeatmapCellColorCanvas';
 
 export default function Heatmap(props) {
   const {
-    clusters, selectedCellIds, cellColors, clearPleaseWait,
+    uuid,
+    cells,
+    clusters,
+    selectedCellIds,
+    cellColors,
+    clearPleaseWait,
+    updateCellsHover = (hoverInfo) => {
+      console.warn(`Heatmap updateCellsHover: ${hoverInfo.cellId}`);
+    },
+    updateStatus = (message) => {
+      console.warn(`Heatmap updateStatus: ${message}`);
+    },
   } = props;
   if (clearPleaseWait && clusters) {
     clearPleaseWait('clusters');
@@ -13,17 +24,29 @@ export default function Heatmap(props) {
   return (
     <React.Fragment>
       <HeatmapCellColorCanvas
+        uuid={uuid}
+        cells={cells}
         clusters={clusters}
         cellColors={cellColors}
+        updateCellsHover={updateCellsHover}
+        updateStatus={updateStatus}
         height="15%"
       />
       <HeatmapCellSelectionCanvas
+        uuid={uuid}
+        cells={cells}
         clusters={clusters}
         selectedCellIds={selectedCellIds}
+        updateCellsHover={updateCellsHover}
+        updateStatus={updateStatus}
         height="15%"
       />
       <HeatmapDataCanvas
+        uuid={uuid}
+        cells={cells}
         clusters={clusters}
+        updateCellsHover={updateCellsHover}
+        updateStatus={updateStatus}
         height="70%"
       />
     </React.Fragment>