+++ src/components/heatmap/HeatmapSubscriber.js
@@ -3,19 +3,25 @@ import PubSub from 'pubsub-js';
 
 import TitleInfo from '../TitleInfo';
 import {
-  CELLS_COLOR, CLUSTERS_ADD, CELLS_SELECTION, CLEAR_PLEASE_WAIT,
+  CELLS_COLOR, CLUSTERS_ADD, CELLS_ADD, CELLS_SELECTION,
+  CLEAR_PLEASE_WAIT, CELLS_HOVER, STATUS_INFO,
 } from '../../events';
 import Heatmap from './Heatmap';
 
 export default class HeatmapSubscriber extends React.Component {
   constructor(props) {
     super(props);
-    this.state = { clusters: null, selectedCellIds: {}, cellColors: null };
+    this.state = {
+      cells: {}, clusters: null, selectedCellIds: new Set(), cellColors: null,
+    };
   }
 
   componentWillMount() {
     this.clustersAddToken = PubSub.subscribe(
       CLUSTERS_ADD, this.clustersAddSubscriber.bind(this),
+    );
+    this.cellsAddToken = PubSub.subscribe(
+      CELLS_ADD, this.cellsAddSubscriber.bind(this),
     );
     this.cellsColorToken = PubSub.subscribe(
       CELLS_COLOR, this.cellsColorSubscriber.bind(this),
@@ -32,12 +38,17 @@ export default class HeatmapSubscriber extends React.Component {
 
   componentWillUnmount() {
     PubSub.unsubscribe(this.clustersAddToken);
+    PubSub.unsubscribe(this.cellsAddToken);
     PubSub.unsubscribe(this.cellsColorToken);
     PubSub.unsubscribe(this.cellsSelectionToken);
   }
 
   clustersAddSubscriber(msg, clusters) {
     this.setState({ clusters });
+  }
+
+  cellsAddSubscriber(msg, cells) {
+    this.setState({ cells });
   }
 
   cellsSelectionSubscriber(msg, cellIds) {
@@ -49,20 +60,28 @@ export default class HeatmapSubscriber extends React.Component {
   }
 
   render() {
-    const { clusters, selectedCellIds, cellColors } = this.state;
+    const {
+      cells, clusters, selectedCellIds, cellColors,
+    } = this.state;
     const cellsCount = clusters ? clusters.cols.length : 0;
     const genesCount = clusters ? clusters.rows.length : 0;
-    const selectedCount = selectedCellIds ? Object.keys(selectedCellIds).length : 0;
+    const selectedCount = selectedCellIds ? selectedCellIds.size : 0;
+    const { children, uuid } = this.props;
     return (
       <TitleInfo
         title="Heatmap"
         info={`${cellsCount} cells × ${genesCount} genes,
                with ${selectedCount} cells selected`}
       >
+        {children}
         <Heatmap
+          uuid={uuid}
+          cells={cells}
           clusters={clusters}
           selectedCellIds={selectedCellIds}
           cellColors={cellColors}
+          updateCellsHover={hoverInfo => PubSub.publish(CELLS_HOVER, hoverInfo)}
+          updateStatus={message => PubSub.publish(STATUS_INFO, message)}
           clearPleaseWait={
             layerName => PubSub.publish(CLEAR_PLEASE_WAIT, layerName)
           }