+++ src/components/heatmap/utils.js
@@ -1,3 +1,5 @@
+import { makeCellStatusMessage } from '../utils';
+
 export function setImageDataRGBA(imageData, offset, r, g, b, a) {
   /* eslint-disable no-param-reassign */
   imageData.data[offset + 0] = r;
@@ -9,4 +11,38 @@ export function setImageDataRGBA(imageData, offset, r, g, b, a) {
 
 export function getImageRendering() {
   return /Chrome/.test(navigator.userAgent) ? 'pixelated' : 'crisp-edges';
+}
+
+export function onHeatmapMouseMove(event, props) {
+  const {
+    uuid,
+    cells,
+    clusters,
+    updateCellsHover = (hoverInfo) => {
+      console.warn(`onHeatmapMouseMove updateCellsHover: ${hoverInfo.cellId}`);
+    },
+    updateStatus = (message) => {
+      console.warn(`onHeatmapMouseMove updateStatus: ${message}`);
+    },
+  } = props;
+  // Compute x position relative to the canvas.
+  const rect = event.target.getBoundingClientRect();
+  const pixelX = (event.clientX - rect.left);
+  const { width } = rect;
+  // Cell columns are not exactly equal to individual pixels,
+  // so need to scale by number of cells.
+  const colX = Math.floor((pixelX / width) * clusters.cols.length);
+  // Use the column x-coordinate to look up the cell ID.
+  const cellId = clusters.cols[colX];
+  if (cellId) {
+    // Use the cell ID to look up the cell information object.
+    const cellInfo = cells[cellId];
+    updateCellsHover({
+      cellId,
+      mappings: { xy: cellInfo.xy, ...cellInfo.mappings },
+      uuid,
+      factors: cellInfo.factors,
+    });
+    updateStatus(makeCellStatusMessage(cellInfo.factors));
+  }
 }