+++ src/components/utils.js
@@ -1,7 +1,13 @@
 import { COORDINATE_SYSTEM } from 'deck.gl';
 import { interpolatePlasma } from 'd3-scale-chromatic';
 
-export function cellLayerDefaultProps(cells, updateStatus) {
+export function makeCellStatusMessage(cellInfoFactors) {
+  return Object.entries(cellInfoFactors).map(
+    ([factor, value]) => `${factor}: ${value}`,
+  ).join('; ');
+}
+
+export function cellLayerDefaultProps(cells, updateStatus, updateCellsHover, uuid) {
   return {
     coordinateSystem: COORDINATE_SYSTEM.IDENTITY,
     data: Object.entries(cells),
@@ -13,11 +19,16 @@ export function cellLayerDefaultProps(cells, updateStatus) {
     getLineWidth: 0,
     onHover: (info) => {
       if (info.object) {
-        updateStatus(
-          Object.entries(info.object[1].factors).map(
-            ([factor, value]) => `${factor}: ${value}`,
-          ).join('; '),
-        );
+        updateStatus(makeCellStatusMessage(info.object[1].factors));
+        updateCellsHover({
+          cellId: info.object[0],
+          mappings: { xy: info.object[1].xy, ...info.object[1].mappings },
+          uuid,
+          factors: info.object[1].factors,
+        });
+      } else {
+        // Clear the currently-hovered cell info by passing null.
+        updateCellsHover(null);
       }
     },
   };